PIL-1715 - Updated telephone validation to catch preceding "++" (#443)
An issue was found where a user was able to input a telephone number
preceding with "++".

I've since improved the regex across all of pillar2 and unified the
telephone regex used where it's possible to input a telephone number.
Some DetailFormProvider's didn't have matching unit tests so these have
been added in too.