PIL-2248: Migrate Below Threshold Notification flow to Pillar2-frontend (#496)
The following UI topic in submissions frontend needs to be included in
Pillar2-frontend:
Below Threshold Notification
With a feature flag - Set to 'Off'.
There will be a code freeze for the section during the move

This has largely been a 'lift and shift' exercise. However, I have
amended the manner in which the Phase2ScreensAction works. Initially
this action was working with a default DataRequest but when I introduced
the BTN pages some of them were using a SubscriptionDataRequest. I opted
to change the action to use an IdentifierRequest as they all have the
same identify step/action in common which allows them all the use the
action without any compatibility issues.

---------

Co-authored-by: JamesMMiller <7511223+JamesMMiller@users.noreply.github.com>
Co-authored-by: Mohammad Dweik <180322631+MoDweik@users.noreply.github.com>