PIL-2292: Update and Align Heading Hierarchy for Phase 2 of the Service(BTN) (#502)
Update the heading structure for the BTN pages in alignment with the
prototypes

All heading structures on the page must follow the correct semantic
hierarchy as seen in the prototype.
The page content, visual layout and styling of the headings must remain
consistent with the prototype.
Heading levels must follow a logical sequence without skipping levels
(e.g., no skipping directly from H1 to H3).

Scenarios can be found here:
https://jira.tools.tax.service.gov.uk/browse/PIL-2292

---------

Co-authored-by: JamesMMiller <7511223+JamesMMiller@users.noreply.github.com>
Co-authored-by: Mohammad Dweik <180322631+MoDweik@users.noreply.github.com>