[PIL-2315] - add check using pattern matching for double clicking button (#508)
This PR handles idempotency when it comes to clicking buttons which send
off some information to ETMP

Our previous solution had a technical issue where even though the
property `preventDoubleClicks` was added to the button, it would still
end up double clicking when network speeds were slow enough (3g speed
and lower)

This solution handles it in a more "backend" style approach where before
the logic that is tied to that button is executed, we will first check
if the task has already been done. If not, we will start the logic again
and if so, we will simply go straight to the redirected page. This
approach means that on double clicks and even on multiclicks, it should
remain far more robust