[PIL-2321] - redirect Top Banner Link to DashboardController (#516)
Currently, when you finish the registration process, when you click the
central link in the banner at the top of the page, it takes you to the
same registration completion page. This defeats the purpose of it. To
fix this, we have supplied the banner url explicitly to the View class
to avoid the orElse logic in the layout

Key Changes:
- redirect top banner link to dashboard when already on confirmation
page
- redirect top banner link to dashboard when already on dashboard page
- unit tests