PIL-2453: resolved set up for CheckYourAnswersViewSpec allowing address tests to use country code correctly (#540)
As part of a different ticket, it was identified that the
CheckYourAnswersView didn't have a spec file. A CheckYourAnswersViewSpec
has been created to try and show at least one valid scenario though
there are multiple possible permutations.

The main issue at this stage is that two tests (for addresses) are
failing due to the country code not being used successfully to retrieve
the country name. These two test lines have been commented out in the
view spec. Fix the spec and make sure all the tests work in this spec.