Revert "PIL-2431 Address hover feedback from a11y report (#548)" (#551)
This reverts commit 627945963f43bcf9e2798cb61ffff97e3af999c3 (#548)

We aren't sure that
- The button highlighting is significant enough for low vision users
- The on-hover behaviour where one radio is already selected doesn't
represent an a11y regression, since it's possible to create a scenario
where it appears that 2 radios are focused at the same time.

While I work out the kinks and pin down exactly what we want to achieve
here, I want to roll this back to avoid blocking other releases.