PIL-2478: Clicking top banner link from homepage linked pages should return user to homepage, not registration confirmation page (#547)
For each of the links on the new Homepage, when they are clicked and the
user goes to the respective pages, then ensure that when the user clicks
the top banner link from the page they have been redirected to, that
they are returned to the new Homepage
Due and Overdue page returns user back to the New Homepage
Submission History page returns user back to the New Homepage
Transaction History page returns user back to the New Homepage
Outstanding payment page returns user back to the New Homepage
Request a repayment page returns user back to the New Homepage
Request a repayment page returns user back to the New Homepage
Contact details page returns user back to the New Homepage
Group details page returns user back to the New Homepage
Below-Threshold Notification’ page returns user back to the New Homepage

Each scenario found here:
https://jira.tools.tax.service.gov.uk/browse/PIL-2478

Additionally, I corrected the BTNWaitingRoomView
(/below-threshold-notification/processing-submission) to have a matching
title and heading "Submitting your Below-Threshold Notification" -
confirmed with Lindsay.

Update:
Following some concerns around the broader fix and potential knock-on
effects so close to go live, we've opted to go with the simple fix of
manually updating the banner url per impacted page. This has turned out
to be more than just the 8 pages addressed in the body of the ticket.
This has basically covered each section that is linked to from the
homepage (such as the full BTN journey and the manage account contact
details pages). I've also introduced multiple missing view specs to
ensure we have sufficient unit tests in place.