PIL-2480 Introduce text wrapping for all headings (#552)
A number of screens were identified where user-inputted values are used
in the page heading and text wrapping was not in place to handle long
values. This PR styles all h1 elements to wrap when the content is too
long, and removes the long-word class previously used. It was only
applied on h1 elements, so this is a safe change.

https://jira.tools.tax.service.gov.uk/browse/PIL-2480