PIL-2366 Fixed broken test (#561)
There were recent changes to the way we use Reverse Routing in tests
from `controllers.routes.DashboardController.onPageLoad.url` to
`routes.DashboardController.onPageLoad.url` and one of the tests was
missing the `import controllers.routes`.

The PR builder for #550 failed to detect the issue and deployment
failed.

This PR rectifies this issue.