PIL-2449 New DNA content and domain model for financial data (#563)
The requirement in
[PIL-2449](https://jira.tools.tax.service.gov.uk/browse/PIL-2449) to add
additional states to the dynamic notification area requires nontrivial
parsing of the transactions we receive from ETMP. This PR does that,
alongside introducing a domain model for the financial data, as trying
to understand the new states alongside the old ones has significant
complexity.

- Enumerate ETMP references we understand.
- Model financial transactions as an ADT to support clearer use later
on.
- Move validation of outstanding charges from Service -> Controller to
Connector -> Service. This does not introduce any extra checks.
- Drop any transactions which do not pass validation. This preserves the
existing behaviour.
- Add new content to the dynamic notification area.