PIL-2519 Added logic to prevent MultiNationalEntity from changing to Domestic (#575)
Refactored the logic of the "Manage Group Details" -> "Change" ->
"Entity Locations" form so that a MultiNational Entity (`UkAndOther` or
MNE) is not allowed to change its location to Domestic (`Uk`).

PR #573 has already implemented the changed view on the "Entity
Locations" form/page and also the page that a user is redirected when
trying to change from `UkAndOther` to `Uk`.


Allowed scenarios:
- Domestic -> Domestic
- Domestic -> MultiNational Entity (MNE)
- MultiNational Entity (MNE) -> MultiNational Entity (MNE)

Disallowed scenario:
- MultiNational Entity (MNE) -> Domestic

---

I renamed views, messages and controllers from `MttToDtt` to
`MneToDomestic` to match preexisting names used.