PIL-2553: Added Accounting period text to due and overdue payments page and submission history page (#579)
Given a page displays a table with accounting period dates e.g.
Outstanding Payments
Submission History
Due and Overdue Returns (already done)
 
When that page is selected,
Then the accounting period must be displayed clearly in the following
format:
"Accounting period: 28 September 2023 to 27 September 2024"
 
The text in the example above needs to be written in the correct header
format (bold).