NOJIRA bump sbt memory to 2GB (#585)
When compiling tests from scratch, I am seeing GC errors, like
`[warn] In the last 10 seconds, 9.72 (97.8%) were spent in GC. [Heap:
0.00GB free of 1.00GB, max 1.00GB] Consider increasing the JVM heap
using '-Xmx' or try a different collector, e.g. '-XX:+UseG1GC', for
better performance.`

G1GC works best with large heaps, and we're only on the default 1GB, so
bumping to 2 shouldn't be a big deal.

Note that I think this was caused by my addition of
`"-Ypatmat-exhaust-depth", "off"` to the compiler flags. My view is that
avoiding the compiler abandoning exhaustivity checks is worth an extra
GB of memory, but I'm happy to discuss.