PIL-2495 Add auditing to BTN duplicate submission action (#587)
Currently, we only issue explicit BTN audit events when we make an API
call to submit. We would like to add an extra event to indicate when a
user has attempted to re-submit a BTN.

This PR adds the new event, along with doing a bit of cleanup in the
existing test suite to speed it up; it was running pretty slowly.

[PIL-2495](https://jira.tools.tax.service.gov.uk/browse/PIL-2495)