PIL-2578 Forward details from BTN response to explicit audit events (#581)
We currently populate the BTN explicit audit events with fixed values,
when they should be using the values we are returned by the pillar2
backend. This PR captures the code and message values from an error
response, and passes them to the audit event call.

Note that the error code & message surfaced by the backend is not
necessarily the same as those returned by ETMP. That is not feasible to
address right now, so this is an improvement more than a total and
complete "fix" for the audit event data.