PIL-2560 Simplify duplication across financial model (#594)
Under #563, I introduced a `TaxPeriod` model with the acknowledgement
that it was very similar to `AccountingPeriod`. This PR revisits that,
and unifies the model, removing `TaxPeriod`. It also makes a couple of
extra tweaks to leverage the typesafety of the newer modelling.

I couldn't find any other areas with obvious opportunities for
simplification, beyond a few controllers that would benefit from a
service layer. I'll stick some bits in the backlog with a very low
priority in case we get round it at some point.

I also removed a bunch of unused models & JSON formats, so it's probably
easiest to work through this one commit at a time.