PIL-2196 Introduce generic spinning wheel component (#598)
Moved here from #596 (I had used the wrong branch name).

We have various spinning wheels across the service. Each was implemented
separately. This PR introduces a spinning wheel component which can be
reused across each waiting room. Following that, we introduce a generic
waiting room route, with typesafe configuration controlling routing for
each scenario (provided the resulting non-exhaustive match warning is
addressed, since we don't have sbt-tpolecat here).

I'm not totally sold on the `LongRunningSubmission` naming, but I
couldn't come up with anything better. Very much open to suggestions.

Ideally, I'd like the interface to support submitting a "task" and then
being able to look up the state on refresh. However, that would have
made this PR even larger than it already is, so I left the existing
persistence as-is, instead opting for an interface that can look up each
kind of submission. I think there's scope to expand on this if/when we
try to extract some logic out of the CYA controllers and into
corresponding services, though it's not a must.