[PIL-2599] - Fix infinite loading bug (#601)
Right now there is a bug where when a user registers a new org and
creates the subscription in a stubbed env. After this they go to the
homepage then they go to manage group details. If they go into this
journey and then submit changes, they will get an infinite waiting room
loop. This ticket is to show an error screen instead of infinitely
loading as that is more appropriate user feedback

Key Changes
- Refactored for comprehension to break properly when reference number
is not found.

---------

Co-authored-by: dc-smith <75583950+dc-smith@users.noreply.github.com>