PIL-2625 track submission time in RFM & BTN journeys (#603)
We currently generate a submission timestamp for the RFM & BTN
confirmation screens when the page loads. This is not correct, because
it means refreshing the pages can lead to changed (and inaccurate)
timestamps.

This PR persists the point at which we deem a submission "successful,"
and then uses it to render the confirmation pages.
I opted for this over persisting the timestamp at the point we make the
request to avoid the possibility of submitting, having it fail, and then
manually navigating to the confirmation URL.