PIL-2609 ensure InProgress writes first when submitting a repayment (#609)
We've seen cases of the acceptance tests getting stuck in the waiting
room when attempting to submit a repayment. I don't have any direct
proof, but I think this is a possible cause.

Currently, we schedule a future to write InProgress to mongo. We then
immediately schedule a second which makes the repayment request, handles
the result, and writes back. We wait for neither to complete, and
respond with a redirect immediately.

This PR changes things, so that the InProgress write must succeed before
we redirect, and (more importantly), it must succeed _before_ we make
the repayment request. I suspect the ATs have been hitting cases where
the submission request completes first, writes the result back, and then
the initial InProgress write goes in over the top.

The browser then loads the waiting room, reading an InProgress state,
and getting stuck.

Again, no proof that this is the cause of the failing ATs, though I
think this is worth fixing regardless. As with other waiting
room-adjacent stuff, this could all do with a larger refactor in future.